refactor : 엔티티 변경 및 dto 수정
reward
- reward_withdrawal_account 엔티티 관계 x -> reward_weithdrawal 엔티티 관계 OneToOne
- dto :
   - RewardDto
     - RewardType 속성 추가
     - RewardService mapRewardToDto 수정

reward_withdrawal
- withdrawal_status 속성 추가
  -> 관련 dto에 해당 속성 추가 withdrawalHistoryDto

reward_account
- reward 관계 삭제

product
- rate 관련 속성 int -> float
- int 타입 -> Integer
- DTO
  - ProductCategoryDto
    - product id -> productCode
 - ProductOptionDTO
   - value : List<String> -> String : 굳이 리스트로 한번 묶을 필요가 없어서 지움

+++ src/main/java/com/freeder/buclserver/domain/product/entity/Product.java
@@ -81,17 +81,17 @@ public class Product extends TimestampMixin implements Serializable {
 	private TaxStatus taxStatus;
 
 	@Column(name = "margin_rate")
-	private int marginRate;
+	private Integer marginRate;
 	@Column(name = "tax_rate")
-	private int taxRate;
+	private Integer taxRate;
 	@Column(name = "discount_rate")
-	private int discountRate;
+	private Float discountRate;
 
 	@Column(name = "consumer_reward_rate")
-	private int consumerRewardRate;
+	private Float consumerRewardRate;
 
 	@Column(name = "business_reward_rate")
-	private int businessRewardRate;
+	private Float businessRewardRate;
 
 	@Column(name = "image_path", length = 5000)
 	private String imagePath;
@@ -107,7 +107,7 @@ public class Product extends TimestampMixin implements Serializable {
 	private String skuCode;
 
 	@Column(name = "product_priority")
-	private int productPriority;
+	private Integer productPriority;
 
 	@ColumnDefault("true")
 	@Column(name = "is_exposed")
